Rename custom_cops_generator to rubocop-extension-generator
