+++ Gemfile
@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in custom_cops_generator.gemspec
+# Specify your gem's dependencies in rubocop-extension-generator.gemspec
 gemspec
 
 gem "rake", "~> 12.0"
