--- exe/custom_cops_generator
@@ -1,5 +1,0 @@
-#!ruby
-
-require 'custom_cops_generator'
-
-CustomCopsGenerator::CLI.run(ARGV)
