--- lib/custom_cops_generator/version.rb
@@ -1,3 +1,0 @@
-module CustomCopsGenerator
-  VERSION = "0.1.2"
-end
