+++ lib/rubocop/extension/generator.rb
@@ -1,0 +1,17 @@
+require "rubocop/extension/generator/version"
+require "rubocop/extension/generator/cli"
+require "rubocop/extension/generator/generator"
+
+require 'active_support'
+require 'active_support/core_ext/string/inflections'
+
+require 'optparse'
+require 'pathname'
+require 'fileutils'
+
+module RuboCop
+  module Extension
+    module Generator
+    end
+  end
+end
