+++ lib/rubocop/extension/generator/version.rb
@@ -1,0 +1,7 @@
+module RuboCop
+  module Extension
+    module Generator
+      VERSION = "0.1.2"
+    end
+  end
+end
